agentcore/access_token.py
"""
Access Token Management for AgentCore Gateway
Handles OAuth token retrieval from Cognito (if needed)
Simplified version for basic setup
"""
import os
import requests
from dotenv import load_dotenv

load_dotenv()

def get_gateway_access_token():
    """
    Get access token for Gateway authentication
    For basic setup without Cognito, returns None
    Can be extended for OAuth if needed
    """
    # For basic setup, we may not need Cognito
    # Return None or a simple token if authentication is not required
    bearer_token = os.getenv("BEARER_TOKEN")
    
    if bearer_token:
        return bearer_token
    
    # If Cognito is configured, implement OAuth flow here
    cognito_domain = os.getenv("COGNITO_DOMAIN")
    cognito_client_id = os.getenv("COGNITO_CLIENT_ID")
    cognito_client_secret = os.getenv("COGNITO_CLIENT_SECRET")
    
    if cognito_domain and cognito_client_id:
        # Implement OAuth token retrieval if needed
        # For now, return None for basic setup
        pass
    
    return None






