agentcore/requirements-runtime.txt
python-dateutil>=2.8.2
python-dotenv>=1.0.0
requests>=2.31.0
boto3==1.39.7
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents>=0.1.0
strands-agents-tools>=0.1.0






